VC/Vidyo: Use BigInteger for extension

+++ vc_vidyo/indico_vc_vidyo/migrations/201502191158_a702ad0bcea_create_extensions_table.py
@@ -20,7 +20,7 @@ def upgrade():
     op.execute(CreateSchema('plugin_vc_vidyo'))
     op.create_table('vidyo_extensions',
                     sa.Column('vc_room_id', sa.Integer(), nullable=False),
-                    sa.Column('extension', sa.Integer(), nullable=True, index=True),
+                    sa.Column('extension', sa.BigInteger(), nullable=True, index=True),
                     sa.Column('owned_by_id', sa.Integer(), nullable=True, index=True),
                     sa.ForeignKeyConstraint(['vc_room_id'], ['events.vc_rooms.id']),
                     sa.PrimaryKeyConstraint('vc_room_id'),
